refactor!: change axis DSL from translational to rotational units
The axis entity now uses roll/pitch/yaw (in degrees) instead of x/y/z
(in metres) to specify joint axis orientation.

This is more intuitive: an empty axis block defaults to the Z-axis,
and other orientations are specified by rotating from that default.

Examples:
- Z-axis (default): `axis do end`
- Y-axis: `axis do roll(~u(-90 degree)) end`
- X-axis: `axis do pitch(~u(90 degree)) end`

The builder converts Euler angles to a unit vector by applying the
rotation to the default Z-axis {0, 0, 1}.